+++ app/api/chart/v1/router.py
@@ -13,7 +13,7 @@ from app.module.asset.enum import AssetType, MarketIndex
 from app.module.asset.repository.asset_repository import AssetRepository
 from app.module.asset.schema import MarketIndexData, StockAssetSchema
 from app.module.asset.services.asset.asset_query import AssetQuery
-from app.module.asset.services.asset_service import AssetService
+from app.module.asset.services.asset.asset_service import AssetService
 from app.module.asset.services.dividend_service import DividendService
 from app.module.asset.services.realtime_index_service import RealtimeIndexService
 from app.module.auth.constant import DUMMY_USER_ID
