+++ app/module/asset/dependencies/asset_stock_dependency.py
@@ -1,5 +1,5 @@
 from app.module.asset.services.asset_stock.asset_stock_validate import AssetStockValidate
-from app.module.asset.services.asset_stock_service import AssetStockService
+from app.module.asset.services.asset_stock.asset_stock_service import AssetStockService
 from app.module.asset.services.exchange_rate_service import ExchangeRateService
 
 exchange_rate_service = ExchangeRateService()
