+++ app/module/asset/services/asset_stock/asset_stock_service.py
