+++ app/module/chart/dependencies/rich_dependency.py
@@ -1,6 +1,6 @@
 from app.module.asset.services.asset.asset_query import AssetQuery
-from app.module.asset.services.asset_service import AssetService
-from app.module.asset.services.asset_stock_service import AssetStockService
+from app.module.asset.services.asset.asset_service import AssetService
+from app.module.asset.services.asset_stock.asset_stock_service import AssetStockService
 from app.module.asset.services.dividend_service import DividendService
 from app.module.asset.services.exchange_rate_service import ExchangeRateService
 from app.module.asset.services.index_daily_service import IndexDailyService
@@ -8,7 +8,7 @@ from app.module.asset.services.index_minutely_service import IndexMinutelyServic
 from app.module.asset.services.realtime_index_service import RealtimeIndexService
 from app.module.asset.services.stock_daily_service import StockDailyService
 from app.module.asset.services.stock_minutely_service import StockMinutelyService
-from app.module.asset.services.stock_service import StockService
+from app.module.asset.services.stock.stock_service import StockService
 from app.module.chart.services.rich_service import RichService
 
 exchange_rate_service = ExchangeRateService()
