+++ app/module/chart/dependencies/summary_dependency.py
@@ -1,9 +1,9 @@
-from app.module.asset.services.asset_service import AssetService
-from app.module.asset.services.asset_stock_service import AssetStockService
+from app.module.asset.services.asset.asset_service import AssetService
+from app.module.asset.services.asset_stock.asset_stock_service import AssetStockService
 from app.module.asset.services.dividend_service import DividendService
 from app.module.asset.services.exchange_rate_service import ExchangeRateService
 from app.module.asset.services.stock_daily_service import StockDailyService
-from app.module.asset.services.stock_service import StockService
+from app.module.asset.services.stock.stock_service import StockService
 from app.module.chart.services.summary_service import SummaryService
 
 stock_daily_service = StockDailyService()
